chore: refactor BlueBubbles clients and models
