+++ app/logger.py
@@ -1,0 +1,5 @@
+"""Logger configuration."""
+
+import structlog
+
+log: structlog.stdlib.BoundLogger = structlog.get_logger()
